Add support for assuming a roleARN for cloudwatch.
This adds a flag to buildkite-agent-metrics that is the roleARN
that should be assumed when talking to the CW API.

Once we establish the proper role in the ci account we'll configure
the treb task definition to use this role.  Hopefully this should
allow the production metrics container to emit to the ci CW environment.